requirements.txt
pypdf>=3.17.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
reportlab>=4.0.0
langchain>=0.0.200
transformers>=4.30.0
numpy>=1.24.0
torch>=2.0.0
tqdm>=4.65.0
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
sentencepiece>=0.1.99
